Remove post-link script on Windows
For some reason, this script fails on some configurations with
exit code 1. This is surprising, since all it does is echoing
into the message file, and this message gets displayed correctly.
Since the message is not that important, and failing with an error
during installation *is*, let's just remove it.